Fixup manager cleanup and waiting
