--- examples/clone_vs_add_hook/utils.go
@@ -1,55 +1,0 @@
-package main
-
-import (
-	"encoding/binary"
-	"log"
-	"os"
-	"time"
-	"unsafe"
-)
-
-// ByteOrder - host byte order
-var ByteOrder binary.ByteOrder
-
-func init() {
-	ByteOrder = getHostByteOrder()
-}
-
-// getHostByteOrder - Returns the host byte order
-func getHostByteOrder() binary.ByteOrder {
-	var i int32 = 0x01020304
-	u := unsafe.Pointer(&i)
-	pb := (*byte)(u)
-	b := *pb
-	if b == 0x04 {
-		return binary.LittleEndian
-	}
-
-	return binary.BigEndian
-}
-
-// trigger - Creates and then removes a tmp folder to trigger the probes
-func trigger() error {
-	log.Println("Generating events to trigger the probes ...")
-	// Creating a tmp directory to trigger the probes
-	tmpDir := "/tmp/test_folder"
-	log.Printf("creating %v", tmpDir)
-	err := os.MkdirAll(tmpDir, 0666)
-	if err != nil {
-		return err
-	}
-
-	// Sleep a bit to give time to the perf event
-	time.Sleep(500 * time.Millisecond)
-
-	// Removing a tmp directory to trigger the probes
-	log.Printf("removing %v", tmpDir)
-	err = os.RemoveAll(tmpDir)
-	if err != nil {
-		return err
-	}
-
-	// Sleep a bit to give time to the perf event
-	time.Sleep(500 * time.Millisecond)
-	return nil
-}
