+++ examples/mapspec_editor/main.go
@@ -3,6 +3,7 @@ package main
 import (
 	"bytes"
 	_ "embed"
+	"fmt"
 	"log"
 	"math"
 
@@ -18,6 +19,12 @@ var Probe []byte
 var m = &manager.Manager{}
 
 func main() {
+	if err := run(); err != nil {
+		log.Fatal(err)
+	}
+}
+
+func run() error {
 	options := manager.Options{
 		MapSpecEditors: map[string]manager.MapSpecEditor{
 			"cache": {
@@ -32,18 +39,18 @@ func main() {
 		},
 	}
 
-	// Initialize the manager
 	if err := m.InitWithOptions(bytes.NewReader(Probe), options); err != nil {
-		log.Fatal(err)
+		return err
 	}
+	defer func() {
+		if err := m.Stop(manager.CleanAll); err != nil {
+			log.Print(err)
+		}
+	}()
 
 	log.Println("successfully loaded, checkout the parameters of the map \"cache\" using bpftool")
 	log.Println("=> You should see MaxEntries 1000000 instead of 10")
+	_, _ = fmt.Scanln()
 
-	wait()
-
-	// Close the manager
-	if err := m.Stop(manager.CleanAll); err != nil {
-		log.Fatal(err)
-	}
+	return nil
 }
