--- examples/programs/perf_event/utils.go
@@ -1,6 +1,0 @@
-package main
-
-// trigger asd
-func trigger() error {
-	return nil
-}
