+++ examples/programs/tc/main.go
@@ -5,9 +5,7 @@ import (
 	_ "embed"
 	"fmt"
 	"log"
-	"os"
-	"os/signal"
-	"syscall"
+	"net/http"
 
 	"github.com/vishvananda/netlink"
 	"golang.org/x/sys/unix"
@@ -42,32 +40,35 @@ var m = &manager.Manager{
 }
 
 func main() {
-	// Initialize the manager
-	if err := m.Init(bytes.NewReader(Probe)); err != nil {
+	if err := run(); err != nil {
 		log.Fatal(err)
 	}
+}
 
-	// Start the manager
+func run() error {
+	if err := m.Init(bytes.NewReader(Probe)); err != nil {
+		return err
+	}
+	defer func() {
+		if err := m.Stop(manager.CleanAll); err != nil {
+			log.Print(err)
+		}
+	}()
 	if err := m.Start(); err != nil {
-		log.Fatal(err)
+		return err
 	}
 
 	log.Println("successfully started, head over to /sys/kernel/debug/tracing/trace_pipe")
 
 	// Generate some network traffic to trigger the probe
 	trigger()
-	wait()
+	_, _ = fmt.Scanln()
 
-	// Close the manager
-	if err := m.Stop(manager.CleanAll); err != nil {
-		log.Fatal(err)
-	}
+	return nil
 }
 
-// wait - Waits until an interrupt or kill signal is sent
-func wait() {
-	sig := make(chan os.Signal, 1)
-	signal.Notify(sig, os.Interrupt, syscall.SIGTERM)
-	<-sig
-	fmt.Println()
+// trigger - Generate some network traffic to trigger the probe
+func trigger() {
+	log.Println("Generating some network traffic to trigger the probes ...")
+	_, _ = http.Get("https://www.google.com/")
 }
