+++ examples/programs/uprobe/main.go
@@ -3,7 +3,10 @@ package main
 import (
 	"bytes"
 	_ "embed"
+	"io"
 	"log"
+	"os/exec"
+	"time"
 
 	manager "github.com/DataDog/ebpf-manager"
 )
@@ -23,14 +26,22 @@ var m = &manager.Manager{
 }
 
 func main() {
-	// Initialize the manager
-	if err := m.Init(bytes.NewReader(Probe)); err != nil {
+	if err := run(); err != nil {
 		log.Fatal(err)
 	}
+}
 
-	// Start the manager
+func run() error {
+	if err := m.Init(bytes.NewReader(Probe)); err != nil {
+		return err
+	}
+	defer func() {
+		if err := m.Stop(manager.CleanAll); err != nil {
+			log.Print(err)
+		}
+	}()
 	if err := m.Start(); err != nil {
-		log.Fatal(err)
+		return err
 	}
 
 	log.Println("successfully started, head over to /sys/kernel/debug/tracing/trace_pipe")
@@ -39,9 +50,23 @@ func main() {
 	if err := trigger(); err != nil {
 		log.Print(err)
 	}
+	return nil
+}
 
-	// Close the manager
-	if err := m.Stop(manager.CleanAll); err != nil {
-		log.Fatal(err)
+// trigger - Spawn a bash and execute a command to trigger the probe
+func trigger() error {
+	log.Println("Spawning a shell and executing `id` to trigger the probe ...")
+	cmd := exec.Command("/usr/bin/bash", "-i")
+	stdinPipe, _ := cmd.StdinPipe()
+	go func() {
+		_, _ = io.WriteString(stdinPipe, "id")
+		time.Sleep(100 * time.Millisecond)
+		_ = stdinPipe.Close()
+	}()
+	b, err := cmd.Output()
+	if err != nil {
+		return err
 	}
+	log.Printf("from bash: %v", string(b))
+	return nil
 }
