+++ examples/programs/xdp/main.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	_ "embed"
 	"log"
+	"net/http"
 
 	manager "github.com/DataDog/ebpf-manager"
 )
@@ -24,23 +25,33 @@ var m = &manager.Manager{
 }
 
 func main() {
-	// Initialize the manager
-	if err := m.Init(bytes.NewReader(Probe)); err != nil {
+	if err := run(); err != nil {
 		log.Fatal(err)
 	}
+}
 
-	// Start the manager
+func run() error {
+	if err := m.Init(bytes.NewReader(Probe)); err != nil {
+		return err
+	}
+	defer func() {
+		if err := m.Stop(manager.CleanAll); err != nil {
+			log.Print(err)
+		}
+	}()
 	if err := m.Start(); err != nil {
-		log.Fatal(err)
+		return err
 	}
 
 	log.Println("successfully started, head over to /sys/kernel/debug/tracing/trace_pipe")
 
 	// Generate some network traffic to trigger the probe
 	trigger()
+	return nil
+}
 
-	// Close the manager
-	if err := m.Stop(manager.CleanAll); err != nil {
-		log.Fatal(err)
-	}
+// trigger - Generate some network traffic to trigger the probe
+func trigger() {
+	log.Println("Generating some network traffic to trigger the probes ...")
+	_, _ = http.Get("https://www.google.com/")
 }
