+++ examples/tests_and_benchmarks/main.go
@@ -33,16 +33,26 @@ var testData = []TestData{
 }
 
 func main() {
-	// Initialize the manager
-	var m = &manager.Manager{}
-	if err := m.Init(bytes.NewReader(Probe)); err != nil {
+	if err := run(); err != nil {
 		log.Fatal(err)
 	}
+}
+
+func run() error {
+	var m = &manager.Manager{}
+	if err := m.Init(bytes.NewReader(Probe)); err != nil {
+		return err
+	}
+	defer func() {
+		if err := m.Stop(manager.CleanAll); err != nil {
+			log.Print(err)
+		}
+	}()
 
 	// Get map used to send tests
 	testMap, found, err := m.GetMap("my_func_test_data")
 	if !found || err != nil {
-		log.Fatalf("couldn't retrieve my_func_test_data %v", err)
+		return fmt.Errorf("couldn't retrieve my_func_test_data %v", err)
 	}
 
 	// Get xdp program used to trigger the tests
@@ -52,29 +62,31 @@ func main() {
 		},
 	)
 	if !found || err != nil {
-		log.Fatalf("couldn't retrieve my_func_test %v", err)
+		return fmt.Errorf("couldn't retrieve my_func_test %v", err)
 	}
 	testProg := testProgs[0]
 
-	// Run test
-	runtTest(testMap, testProg)
-
-	// Run benchmark
-	runtBenchmark(testMap, testProg)
+	if err := runtTest(testMap, testProg); err != nil {
+		return err
+	}
+	if err := runtBenchmark(testMap, testProg); err != nil {
+		return err
+	}
+	return nil
 }
 
-func runtTest(testMap *ebpf.Map, testProg *ebpf.Program) {
+func runtTest(testMap *ebpf.Map, testProg *ebpf.Program) error {
 	log.Println("Running tests ...")
 	for _, data := range testData {
 		// insert data
 		if err := testMap.Put(testDataKey, data); err != nil {
-			log.Fatal(err)
+			return err
 		}
 
 		// Trigger test - (the 14 bytes is for the minimum packet size required to test an XDP program)
 		outLen, _, err := testProg.Test(make([]byte, 14))
 		if err != nil {
-			log.Fatal(err)
+			return err
 		}
 		if data.Input == 42 && data.Output == 128 {
 			log.Printf("(failure expected on next test)")
@@ -85,20 +97,21 @@ func runtTest(testMap *ebpf.Map, testProg *ebpf.Program) {
 			log.Printf("%v - FAIL (checkout /sys/kernel/debug/tracing/trace_pipe to see the logs)", data)
 		}
 	}
+	return nil
 }
 
-func runtBenchmark(testMap *ebpf.Map, testProg *ebpf.Program) {
+func runtBenchmark(testMap *ebpf.Map, testProg *ebpf.Program) error {
 	log.Println("Running benchmark ...")
 	for _, data := range testData {
 		// insert data
 		if err := testMap.Put(testDataKey, data); err != nil {
-			log.Fatal(err)
+			return err
 		}
 
 		// Trigger test
 		outLen, duration, err := testProg.Benchmark(make([]byte, 14), 1000, nil)
 		if err != nil {
-			log.Fatal(err)
+			return err
 		}
 		if data.Input == 42 && data.Output == 128 {
 			log.Printf("(failure expected on next benchmark)")
@@ -109,4 +122,5 @@ func runtBenchmark(testMap *ebpf.Map, testProg *ebpf.Program) {
 			log.Printf("%v - benchmark FAILED (checkout /sys/kernel/debug/tracing/trace_pipe to see the logs)", data)
 		}
 	}
+	return nil
 }
